requirements.txt
pydantic==1.8.2
pyyaml==6.0
jinja2==3.0.2
